move InfoBloc from setting to main_screen
