+++ lib/screens/main_screen/bloc/info_event.dart
@@ -1,0 +1,10 @@
+part of 'info_bloc.dart';
+
+abstract class InfoEvent extends Equatable {
+  const InfoEvent();
+
+  @override
+  List<Object> get props => [];
+}
+
+class LoadInfo extends InfoEvent {}
